I added the following scripts to the masterscript:     --pathlinker-no-div   - PathLinker without the large sum division     --pagerank-pathlinker - RWR+PathLinker     --cyclinker           - CycLinker     --pagerank-cyclinker  - RWR+CycLinker
I modified the calls to PCSF and IPA to call the scripts from SVN.
I added an option --rev-pathways to run the algorithms on the pathways in the reverse order. This is helpful for running on 2 computers
Some algorithms didn't have the allnetpath option so I added those
Finally I added the checkDir command for the viz dir.


git-svn-id: file:///home/murali/svnrepo/group/anna-ritz/projects/2015-03-pathlinker@11733 1fbec0df-f6b4-4bca-857c-de90862b47bd