Add subset analysis & shortcut alg, fix fold calc. This commit includes the following: Subset analysis: prints table of statistics for each pathway to show just how much of the pathway is included in the interactome
Shortcut alg: add algorithm object for the shortcuts algorithm by
Navlakha et. al to the pipeline

Fold calc: there were (and still are) issues with precision/recall. This
commit fixes a non-deterministic bug having to do with taking a set and
casting it to a list, and makes it so that negatives are divided up into
folds as well as positives, and each negative will only count toward the
calculation of precision/recall in a single fold.