Adjust prec/rec calc. for concatenated QuickLinker Prevously, precision/recall curves for an algorithm were aggregated over folds and pathways, using the rank of predictions both times. This adjusts things so that the fold results are themselves aggregated, and then precision/recall is aggregated on the the aggregated fold results. In this manner, if fold2 had a higher scoring prediction for its first prediction than fold1, this did not matter: they were both aggregated with a rank of 1. Now, fold2's item is in a separate rank, followed by fold1's item.
This commit includes necessary adjustments to the precision/recall
pipeline to make this happen. A couple of bugs from the previous
precision/recall calculation have been eliminated.

In tandem, this commit adds the QuickLinkerConcat algorithm for
combining the results of multiple regexes in a row, with preference
and higher rank giving from the first regex to the last regex.